Naive implementation of prop `currVelocity` for both `Spring`s.
When testing I tried to implement a simple "pulse" button, which would
set the velocity for one frame. To do this I used a flag that I would
set to `true`, and immediately back to `false` (in `componentDidUpdate`)
Due to the fact that the `raf` is asynchronous, the owner had time to
render twice, and therefore call `componentWillReceiveProps` twice
and finally setting `this.forceUpdateCurrVelocity` twice. This meant
that the `Spring` would never see the right
`this.forceUpdateCurrVelocity` because it would get overwritten.